+++ addons/mail/wizard/mail_compose_message_views.xml
@@ -19,6 +19,7 @@
                         <field name="lang" invisible="1"/>
                         <field name="mail_server_id" invisible="1"/>
                         <field name="model" invisible="1"/>
+                        <field name="model_is_thread" invisible="1"/>
                         <field name="parent_id" invisible="1"/>
                         <field name="record_name" invisible="1"/>
                         <field name="render_model" invisible="1"/>
