+++ addons/mass_mailing/wizard/mail_compose_message.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class MailComposeMessage(models.TransientModel):
@@ -11,13 +11,6 @@ class MailComposeMessage(models.TransientModel):
     campaign_id = fields.Many2one('utm.campaign', string='Mass Mailing Campaign', ondelete='set null')
     mass_mailing_name = fields.Char(string='Mass Mailing Name', help='If set, a mass mailing will be created so that you can track its results in the Email Marketing app.')
     mailing_list_ids = fields.Many2many('mailing.list', string='Mailing List')
-    model_is_thread = fields.Boolean(compute='_compute_model_is_thread')
-
-    @api.depends('model')
-    def _compute_model_is_thread(self):
-        for composer in self:
-            model = self.env['ir.model']._get(composer.model)
-            composer.model_is_thread = model.is_mail_thread
 
     def _action_send_mail(self, auto_commit=False):
         """ Override to generate the mass mailing in case only the name was
